improve relationsDialog.html manage url params in mainController.js
