+++ public/vocables/index.html
@@ -104,7 +104,7 @@
 
         <script src="js/tools/KGcreator.js"></script>
         <script src="js/tools/KGpropertyFilter.js"></script>
-        <script src="js/tools/compose.js"></script>
+        <script src="js/tools/composer.js"></script>
         <script src="js/tools/sparqlQueryUI.js"></script>
 
 
@@ -230,6 +230,7 @@
             });
 
             $(function () {
+
                 MainController.initConfig(function (err) {
                     // sentry/glitchtip
                     if (Config.sentryDsnJsFront) {
@@ -281,6 +282,7 @@
                     });
 
                     MainController.UI.setCredits();
+
                 });
             });
         </script>
@@ -480,16 +482,13 @@
         </style>
     </head>
     <body>
-        <div class="rowXXX" style="display: flex; flex-direction: column">
+        <div class="rowXXX" style="display: flex; flex-direction: column;    min-height: 30px;">
             <div style="display: flex; flex-direction: row; justify-content: space-evenly">
-                <!--</div> <div style="display: flex;flex-direction: row">-->
-                <!-- <div class="colXXX" style="min-height: 25px;"-->
+
                 <div>
                     <table>
                         <tr>
-                            <td>
-                                <!-- <button onclick="MainController.test()">T</button>-->
-                            </td>
+
                             <td><img src="./images/waitAnimated.gif" id="waitImg" style="display: none; width: 15px; text-align: justify" /></td>
                             <td><span class="messageDiv" id="messageDiv"></span></td>
                         </tr>
