+++ public/vocables/js/tools/lineage/lineage_containers.js
@@ -700,7 +700,7 @@ var Lineage_containers = (function () {
                             filter: Sparql_common.setFilter("prop", objectProperties),
                         };
                         options.allNodes = false;
-                        Lineage_classes.drawRelations(null, null, "Properties", options);
+                        Lineage_relations.drawRelations(null, null, "Properties", options);
                     }
                     return callbackSeries();
                 },
