+++ public/vocables/js/tools/lineage/lineage_properties.js
@@ -261,6 +261,7 @@ action: function (_e) {
             if (err) {
                 return callback(err);
             }
+           result=Lineage_classes.truncateResultToVisGraphLimit(result)
             Sparql_common.setSparqlResultPropertiesLabels(source, result, "prop", function (err, result2) {
                 if (err) {
                     return callback(err);
@@ -268,6 +269,8 @@ action: function (_e) {
                 var visjsData = { nodes: [], edges: [] };
                 var existingNodes = visjsGraph.getExistingIdsMap();
                 var color = Lineage_classes.getSourceColor(source);
+
+
                 result2.forEach(function (item) {
                     if (!existingNodes[item.subject.value]) {
                         existingNodes[item.subject.value] = 1;
@@ -1164,7 +1167,7 @@ action: function (_e) {
                 if (!nodeIds) {
                     options.allNodes = true;
                 }
-                Lineage_classes.drawRelations(null, null, "Properties", options);
+                Lineage_relations.drawRelations(null, null, "Properties", options);
                 // Lineage_properties.drawPredicatesGraph(source, nodeIds, properties);
             } else if (target == "table") {
                 //  Lineage_classes.graphNodeNeighborhood(data, "outcoming", function(err, result) {
