+++ public/vocables/snippets/composer/centralPanel.html
@@ -110,9 +110,9 @@
             <option>matrix</option>
             <option>table</option>
         </select>
-        <button onclick="Compose.resetMatrixPropertiesFilters()">Reset</button>
-        &nbsp;<button onclick="Compose.saveNewRestrictionFilterTriples()">Save new</button> &nbsp;<button onclick="Compose.matrix.exportMatrix()">Export</button> &nbsp;<button
-            onclick="Compose.matrix.copyMatrix()"
+        <button onclick="Composer.resetMatrixPropertiesFilters()">Reset</button>
+        &nbsp;<button onclick="Composer.saveNewRestrictionFilterTriples()">Save new</button> &nbsp;<button onclick="Composer.matrix.exportMatrix()">Export</button> &nbsp;<button
+            onclick="Composer.matrix.copyMatrix()"
         >
             Copy
         </button>
