+++ public/vocables/snippets/composer/leftPanel.html
@@ -14,9 +14,9 @@
 <div style="margin-left: 10px">
     <div>
         Source :
-        <select id="Compose_sourceSelect" onchange="Compose.onChangeSourceSelect($(this).val())"></select>
+        <select id="Compose_sourceSelect" onchange="Composer.onChangeSourceSelect($(this).val())"></select>
     </div>
-    <!--button onclick="Compose.listFiles()">List</button-->
+    <!--button onclick="Composer.listFiles()">List</button-->
     <div>Search : <input id="Compose_searchInPropertiesTreeInput" value="" style="width: 200px; font-size: 12px; margin: 3px; padding: 3px" /></div>
     <div class="Compose_tree_container" style="height: 1000px !important; font-size: 12px">
         <div id="Compose_dataContainerTreeDiv"></div>
