+++ public/vocables/snippets/composer/rightPanel.html
@@ -53,7 +53,7 @@
         <div class="max-height" id="Compose_businessObjects_accordionPanel">
             <div id="Compose_businessObjects_searchDiv">
                 <input id="GenericTools_searchAllSourcesTermInput" style="width: 150px" value="" />
-                <button class="btn btn-sm my-1 py-0 btn-outline-primary" id="GenericTools_onSearchAllInput" onclick="Compose.rightPanelsActions.searchBusinessObjects()">Search</button>
+                <button class="btn btn-sm my-1 py-0 btn-outline-primary" id="GenericTools_onSearchAllInput" onclick="Composer.rightPanelsActions.searchBusinessObjects()">Search</button>
                 <button class="btn btn-sm my-1 py-0 btn-outline-primary" onclick="SourceBrowser.showThesaurusTopConcepts();">reset</button>
                 <button class="btn btn-sm my-1 py-0 btn-outline-primary GenericTools_searchAllOptional" onclick="Export.exportTeeToDataTable(null,'#')">toTable</button>
                 <!--   <button class='btn btn-sm my-1 py-0 btn-outline-primary' onclick="SourceBrowser.exportSearchResult()">CSV</button>-->
